refactor(web-component): move storage key to constants, isolate password manager tests
Move the last-submitted-login-id sessionStorage key next to the other
storage key constants, and move the password manager tests out of the
misc suite into their own descope-wc.password-managers.test.ts.
